msm7x30: Camera HAL for HTC Sense devices (3/2)
This adds an option to the camera to "force" the preview. When using a
wrapper HAL for a GB libcamera, this will fix preview not being
available until a picture is taken.
This option can be default enabled by using a strings.xml overlay
containing <string
name="pref_camera_force_preview_default">off</string> but is provided
as a menu option in the camera itself to allow selection without
rebuilding camera.

Change-Id: Ie72093502a2ad121213b92f5269aab619e5e2f87